+++ RELEASE_NOTES.md
@@ -41,6 +41,7 @@
 - The `Markdown` dependency was bumped so we don't need to add a `type: ignore` due to incorrect type hints.
 - The generated docs now show the symbol type in the table of contents.
 - The dependecies were updated to the latest versions.
+- Disabled some `pylint` checks that are already checked by other tools.
 
 ## Bug Fixes